fix(pipeline): decode VAE in float32 for all render paths
The base render path was decoding VAE latents in bfloat16 while the
inpaint/edit path (fixed in PR #152) used float32. This precision loss
during VAE decode is the primary cause of soft/low-fidelity output from
CivitAI checkpoints. All VAE weights are now loaded as float32, and all
decode paths (base, CivitAI, AIO, control, fallback) use float32.

+++ Sources/ZImage/Pipeline/PipelineUtilities.swift
@@ -94,7 +94,7 @@ public enum PipelineUtilities {
         vae: VAEImageDecoding,
         height: Int,
         width: Int,
-        dtype: DType = .bfloat16
+        dtype: DType = .float32
     ) -> MLXArray {
         let input: MLXArray
         if latents.dtype == dtype {
